labs/cisco-webex/src/maticmind/requirements.txt
backoff==2.2.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
coloredlogs==15.0.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
future==0.18.3
greenlet==1.1.2
humanfriendly==10.0
idna==3.4
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.8.0
python-dotenv==0.20.0
requests==2.31.0
requests-toolbelt==1.0.0
SQLAlchemy==1.4.35
urllib3==2.0.4
webex-bot==0.4.0
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.3.3
zipp==3.8.0
